#914 - Modularize project import/export
- Move export API into API module
