+++ webanno-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/export/ProjectExportService.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.export;
+package de.tudarmstadt.ukp.clarin.webanno.api.export;
 
 import java.io.File;
 import java.util.zip.ZipFile;
