+++ webanno-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/export/ProjectImportRequest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.export;
+package de.tudarmstadt.ukp.clarin.webanno.api.export;
 
 import java.io.Serializable;
 import java.util.Queue;
